pyproject.toml
[project]
name = "distilabel-dataset-generator"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "davidberenstein1957", email = "david.m.berenstein@gmail.com"},
]
dependencies = [
    "distilabel[hf-inference-endpoints,argilla,outlines]>=1.4.1",
    "gradio[oauth]<5.0.0",
    "transformers>=4.44.2",
    "sentence-transformers>=3.2.0",
    "model2vec>=0.2.4",
    "gradio-huggingfacehub-search>=0.0.7",
    "argilla>=2.4.0",
]
requires-python = "<3.13,>=3.10"
readme = "README.md"
license = {text = "apache 2"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
